[FIX] Move agent management to dedicated page
