+++ src/test/App.test.tsx
@@ -10,6 +10,9 @@ vi.mock("../pages/Dashboard", () => ({
 vi.mock("../pages/Vulnerabilities", () => ({
   Vulnerabilities: () => <div>vulnerabilities-page</div>,
 }));
+vi.mock("../pages/Agents", () => ({
+  Agents: () => <div>agents-page</div>,
+}));
 vi.mock("../pages/Users", () => ({ Users: () => <div>users-page</div> }));
 vi.mock("../pages/Settings", () => ({
   Settings: () => <div>settings-page</div>,
@@ -46,6 +49,7 @@ describe("App routes", () => {
   it.each([
     ["/", "dashboard-page"],
     ["/vulnerabilities", "vulnerabilities-page"],
+    ["/agents", "agents-page"],
     ["/users", "users-page"],
     ["/settings", "settings-page"],
     ["/monitoring/scan-1", "dashboard-page"],
