+++ src/test/Settings.test.tsx
@@ -287,148 +287,7 @@ describe("Settings Page", () => {
     });
   });
 
-  it("rotates the agent token and displays it once", async () => {
-    mockState.user.role = "owner";
-    vi.mocked(api.post).mockResolvedValueOnce({
-      data: { agent_token: "ag_rotated-token" },
-    });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(screen.getByRole("button", { name: /Régénérer/i }));
-
-    await waitFor(() => {
-      expect(api.post).toHaveBeenCalledWith("/companies/me/agent-token/rotate");
-      expect(screen.getByText("ag_rotated-token")).toBeInTheDocument();
-      expect(
-        screen.getByText("Nouveau token généré. Copiez-le maintenant."),
-      ).toBeInTheDocument();
-    });
-  });
-
-  it("copies the rotated agent token", async () => {
-    mockState.user.role = "owner";
-    vi.mocked(api.post).mockResolvedValueOnce({
-      data: { agent_token: "ag_rotated-token" },
-    });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(screen.getByRole("button", { name: /Régénérer/i }));
-
-    await screen.findByText("ag_rotated-token");
-    fireEvent.click(screen.getByRole("button", { name: /^Copier$/ }));
-
-    await waitFor(() => {
-      expect(navigator.clipboard.writeText).toHaveBeenCalledWith(
-        "ag_rotated-token",
-      );
-      expect(screen.getByText("Token copié.")).toBeInTheDocument();
-    });
-  });
-
-  it("shows an error when agent token rotation fails", async () => {
-    mockState.user.role = "owner";
-    vi.mocked(api.post).mockRejectedValueOnce({
-      response: { data: { error: "Rotation refusée" } },
-    });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(screen.getByRole("button", { name: /Régénérer/i }));
-
-    await waitFor(() => {
-      expect(screen.getByText("Rotation refusée")).toBeInTheDocument();
-    });
-  });
-
-  it("revokes the agent token", async () => {
-    mockState.user.role = "owner";
-    vi.mocked(api.post).mockResolvedValueOnce({ data: {} });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(screen.getByRole("button", { name: /Révoquer/i }));
-
-    await waitFor(() => {
-      expect(api.post).toHaveBeenCalledWith("/companies/me/agent-token/revoke");
-      expect(screen.getByText("Token agent révoqué.")).toBeInTheDocument();
-    });
-  });
-
-  it("shows an error when agent token revocation fails", async () => {
-    mockState.user.role = "owner";
-    vi.mocked(api.post).mockRejectedValueOnce({
-      response: { data: { error: "Révocation refusée" } },
-    });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(screen.getByRole("button", { name: /Révoquer/i }));
-
-    await waitFor(() => {
-      expect(screen.getByText("Révocation refusée")).toBeInTheDocument();
-    });
-  });
-
-  it("shows client agent management to Aegis personnel only", async () => {
-    mockState.user.role = "admin";
-    vi.mocked(api.get).mockImplementation((url: string) => {
-      if (url === "/admin/companies?search=") {
-        return Promise.resolve({
-          data: [
-            {
-              id: "company-1",
-              name: "Client One",
-              owner_email: "owner@client.test",
-            },
-          ],
-        });
-      }
-      if (url === "/agents?company_id=company-1") {
-        return Promise.resolve({
-          data: {
-            agents: [
-              {
-                id: "agent-1",
-                company_id: "company-1",
-                name: "Production Agent",
-                status: "RUNNING",
-                last_seen: null,
-                created_at: "2026-05-26T09:00:00Z",
-              },
-            ],
-          },
-        });
-      }
-      return Promise.reject(new Error("Not mocked"));
-    });
-
+  it("does not expose agent management tabs anymore", () => {
     render(
       <MemoryRouter>
         <Settings />
@@ -436,149 +295,7 @@ describe("Settings Page", () => {
     );
 
     expect(screen.queryByText("Agents & Token")).not.toBeInTheDocument();
-    fireEvent.click(screen.getByText("Agents clients"));
-
-    expect(
-      await screen.findByText("Agents des entreprises clientes"),
-    ).toBeInTheDocument();
-    expect(await screen.findByText("Production Agent")).toBeInTheDocument();
-    expect(api.get).toHaveBeenCalledWith("/agents?company_id=company-1");
-  });
-
-  it("lets Aegis personnel rotate and revoke a client agent token", async () => {
-    mockState.user.role = "superadmin";
-    vi.mocked(api.get).mockImplementation((url: string) => {
-      if (url === "/admin/companies?search=") {
-        return Promise.resolve({
-          data: [{ id: "company-1", name: "Client One" }],
-        });
-      }
-      if (url === "/agents?company_id=company-1") {
-        return Promise.resolve({ data: { agents: [] } });
-      }
-      return Promise.reject(new Error("Not mocked"));
-    });
-    vi.mocked(api.post)
-      .mockResolvedValueOnce({ data: { agent_token: "ag_client-token" } })
-      .mockResolvedValueOnce({ data: {} });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents clients"));
-    expect((await screen.findAllByText("Client One")).length).toBeGreaterThan(
-      0,
-    );
-    fireEvent.click(screen.getByRole("button", { name: /Régénérer client/i }));
-
-    expect(await screen.findByText("ag_client-token")).toBeInTheDocument();
-    expect(api.post).toHaveBeenCalledWith(
-      "/admin/companies/company-1/agent-token/rotate",
-    );
-
-    fireEvent.click(screen.getByRole("button", { name: /Révoquer client/i }));
-    await waitFor(() => {
-      expect(api.post).toHaveBeenCalledWith(
-        "/admin/companies/company-1/agent-token/revoke",
-      );
-      expect(
-        screen.getByText("Token agent client révoqué."),
-      ).toBeInTheDocument();
-    });
-  });
-
-  it("lets Aegis personnel copy a client token and refresh searched agents", async () => {
-    mockState.user.role = "admin";
-    vi.mocked(api.get).mockImplementation((url: string) => {
-      if (
-        url === "/admin/companies?search=" ||
-        url === "/admin/companies?search=Client"
-      ) {
-        return Promise.resolve({
-          data: [{ id: "company-1", name: "Client One" }],
-        });
-      }
-      if (url === "/agents?company_id=company-1") {
-        return Promise.resolve({ data: { agents: [] } });
-      }
-      return Promise.reject(new Error("Not mocked"));
-    });
-    vi.mocked(api.post).mockResolvedValueOnce({
-      data: { agent_token: "ag_client-copy" },
-    });
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents clients"));
-    await screen.findByText("Token agent client");
-
-    fireEvent.change(
-      screen.getByLabelText("Rechercher une entreprise cliente"),
-      { target: { value: "Client" } },
-    );
-    fireEvent.click(screen.getByRole("button", { name: "Rechercher" }));
-
-    await waitFor(() => {
-      expect(api.get).toHaveBeenCalledWith("/admin/companies?search=Client");
-    });
-
-    fireEvent.click(screen.getByRole("button", { name: "Actualiser" }));
-    fireEvent.click(screen.getByRole("button", { name: /Régénérer client/i }));
-    await screen.findByText("ag_client-copy");
-    fireEvent.click(
-      screen.getByRole("button", { name: /Copier token client/i }),
-    );
-
-    await waitFor(() => {
-      expect(navigator.clipboard.writeText).toHaveBeenCalledWith(
-        "ag_client-copy",
-      );
-      expect(screen.getByText("Token client copié.")).toBeInTheDocument();
-    });
-  });
-
-  it("copies the owner deployment command", async () => {
-    mockState.user.role = "owner";
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
-    fireEvent.click(screen.getByText("Agents & Token"));
-    fireEvent.click(
-      await screen.findByRole("button", { name: /Copier la commande/i }),
-    );
-
-    await waitFor(() => {
-      expect(navigator.clipboard.writeText).toHaveBeenCalledWith(
-        'curl -sL "https://api.aegis-ai.fr/install.sh?token=VOTRE_TOKEN_AGENT" | sudo bash',
-      );
-      expect(
-        screen.getByText("Commande copiée dans le presse-papiers."),
-      ).toBeInTheDocument();
-    });
-  });
-
-  it("does not expose client agent management to a company owner", () => {
-    mockState.user.role = "owner";
-
-    render(
-      <MemoryRouter>
-        <Settings />
-      </MemoryRouter>,
-    );
-
     expect(screen.queryByText("Agents clients")).not.toBeInTheDocument();
-    expect(screen.getByText("Agents & Token")).toBeInTheDocument();
   });
 
   it("shows error when passwords do not match", async () => {
